requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.1
psycopg[binary]
# If you plan to use MySQL instead of PostgreSQL, install:
# mysqlclient==2.2.4


